+++ python-package/lets_plot/__init__.py
@@ -12,6 +12,7 @@ from ._version import __version__
 from ._global_settings import _settings, is_production
 
 from .plot import *
+from .export import *
 from .frontend_context import *
 from .settings_utils import *
 
@@ -19,6 +20,7 @@ __all__ = (plot.__all__ +
            plot_flyer.__all__ +
            frontend_context.__all__ +
            settings_utils.__all__ +
+           export.__all__ +
            ['LetsPlot'])
 
 from .frontend_context import _configuration as cfg